docs: organize Inputs folder
