+++ src/Components/Inputs/Composed/AutoInput.scss
@@ -1,4 +1,4 @@
-@import '../../Styles/modules';
+@import '../../../Styles/modules';
 
 /*
   Auto Input Styles
