+++ src/Components/Inputs/Composed/RangeSlider.scss
@@ -1,4 +1,4 @@
-@import '../../Styles/modules';
+@import '../../../Styles/modules';
 
 $slider-bg: $g2-kevlar;
 $slider-hover-border: $g7-graphite;
