+++ src/index.ts
@@ -50,9 +50,9 @@ export * from './Components/IndexList/IndexListHeader'
 export * from './Components/IndexList/IndexListHeaderCell'
 export * from './Components/IndexList/IndexListRow'
 export * from './Components/IndexList/IndexListRowCell'
-export * from './Components/Inputs/AutoInput'
+export * from './Components/Inputs/Composed/AutoInput'
+export * from './Components/Inputs/Composed/RangeSlider'
 export * from './Components/Inputs/Input'
-export * from './Components/Inputs/RangeSlider'
 export * from './Components/Inputs/TextArea'
 export * from './Components/Label/Label'
 export * from './Components/NavMenu/NavMenu'
